refactor: moved GNSSType enum to api bundle
