+++ kura/org.eclipse.kura.api/src/main/java/org/eclipse/kura/position/GNSSType.java
@@ -1,0 +1,37 @@
+package org.eclipse.kura.position;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public enum GNSSType {
+
+    BEIDOU("Beidou"),
+    GALILEO("Galileo"),
+    GLONASS("Glonass"),
+    GPS("Gps"),
+    MIXED_GNSS_TYPE("MixedGNSSTypes"),
+    OTHER("Other"),
+    UNKNOWN("Unknown");
+
+    private String value;
+
+    private static Map<String, GNSSType> valuesMap = new HashMap<>();
+
+    static {
+        for (GNSSType type : GNSSType.values()) {
+            valuesMap.put(type.getValue(), type);
+        }
+    }
+
+    private GNSSType(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return this.value;
+    }
+
+    public static GNSSType fromValue(String value) {
+        return valuesMap.get(value);
+    }
+}
