+++ kura/org.eclipse.kura.linux.position/src/main/java/org/eclipse/kura/linux/position/PositionServiceImpl.java
@@ -22,6 +22,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.eclipse.kura.KuraErrorCode;
 import org.eclipse.kura.KuraException;
 import org.eclipse.kura.configuration.ConfigurableComponent;
+import org.eclipse.kura.position.GNSSType;
 import org.eclipse.kura.position.NmeaPosition;
 import org.eclipse.kura.position.PositionListener;
 import org.eclipse.kura.position.PositionLockedEvent;
@@ -200,8 +201,8 @@ public class PositionServiceImpl
     }
 
     @Override
-    public String getGnssSystem() {
-        return this.currentProvider.getGnssType().getValue();
+    public GNSSType getGnssType() {
+        return this.currentProvider.getGnssType();
     }
 
     @Override
